requirements.txt
annotated-types==0.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
googlemaps==4.10.0
idna==3.10
munge==1.4.0
phonenumbers==9.0.14
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-whois==0.9.5
PyYAML==6.0.2
requests==2.32.5
six==1.17.0
toml==0.10.2
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
